# Rect.contains() optimization