… src_c/rect.c